defined function that returns whether or not the egg is out of bounds.
note: either use this function to check whether a move will result in the egg being out of bounds or not, or to see ifif it already is. i tried the former, but could not make it work. baka sabog na ko or smth, will retry later. in any case, we can still return to our phase00 out of bounds protection - which is just returning the egg to a valid position every tick.